requirements.txt
astroid==2.1.0
colorama==0.4.1
Django==2.1.7
django-rest-framework==0.1.0
djangorestframework==3.9.1
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.1
pylint==2.2.2
pytz==2018.9
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
